requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.6.1
django-stubs==4.2.7

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Environment management
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.7  # For PostgreSQL support

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
